requirements.txt
streamlit==1.38.0
supabase==2.7.4
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
